[docs] Restore downstream MySQL ModuleID declaration
In the mysql.adoc, restore the downstream comment that specifies the downstream anchor ID for the snapshots topic. I previously removed the comment, because downstream, links to the topic were not working. A tooling change should address this deficit without requiring us to use the upstream anchor IDs.

This change is not visible in either the upstream or downstream content and it has no effect on the upstream behavior.